+++ lib/delayed/exceptions.rb
@@ -3,7 +3,8 @@ require 'timeout'
 module Delayed
   class WorkerTimeout < Timeout::Error
     def message
-      "#{super} (Delayed::Worker.max_run_time is only #{Delayed::Worker.max_run_time.to_i} seconds)"
+      seconds = Delayed::Worker.max_run_time.to_i
+      "#{super} (Delayed::Worker.max_run_time is only #{seconds} second#{seconds == 1 ? '' : 's'})"
     end
   end