+++ lib/delayed/lifecycle.rb
@@ -13,7 +13,9 @@ module Delayed
     }
 
     def initialize
-      @callbacks = EVENTS.keys.inject({}) { |hash, e| hash[e] = Callback.new; hash }
+      @callbacks = EVENTS.keys.each_with_object({}) do |e, hash|
+        hash[e] = Callback.new
+      end
     end
 
     def before(event, &block)
@@ -29,7 +31,7 @@ module Delayed
     end
 
     def run_callbacks(event, *args, &block)
-      missing_callback(event) unless @callbacks.has_key?(event)
+      missing_callback(event) unless @callbacks.key?(event)
 
       unless EVENTS[event].size == args.size
         raise ArgumentError, "Callback #{event} expects #{EVENTS[event].size} parameter(s): #{EVENTS[event].join(', ')}"
@@ -38,17 +40,16 @@ module Delayed
       @callbacks[event].execute(*args, &block)
     end
 
-    private
+  private
 
-      def add(type, event, &block)
-        missing_callback(event) unless @callbacks.has_key?(event)
+    def add(type, event, &block)
+      missing_callback(event) unless @callbacks.key?(event)
+      @callbacks[event].add(type, &block)
+    end
 
-        @callbacks[event].add(type, &block)
-      end
-
-      def missing_callback(event)
-        raise InvalidCallback, "Unknown callback event: #{event}"
-      end
+    def missing_callback(event)
+      raise InvalidCallback, "Unknown callback event: #{event}"
+    end
   end
 
   class Callback