+++ lib/delayed/psych_ext.rb
@@ -1,133 +1,70 @@
 if defined?(ActiveRecord)
   ActiveRecord::Base.class_eval do
+    # rubocop:disable BlockNesting
     if instance_methods.include?(:encode_with)
       def encode_with_override(coder)
         encode_with_without_override(coder)
-        coder.tag = "!ruby/ActiveRecord:#{self.class.name}"
+        coder.tag = "!ruby/ActiveRecord:#{self.class.name}" if coder.respond_to?(:tag=)
       end
       alias_method :encode_with_without_override, :encode_with
       alias_method :encode_with, :encode_with_override
     else
       def encode_with(coder)
-        coder["attributes"] = attributes
-        coder.tag = "!ruby/ActiveRecord:#{self.class.name}"
+        coder['attributes'] = attributes
+        coder.tag = "!ruby/ActiveRecord:#{self.class.name}" if coder.respond_to?(:tag=)
       end
     end
   end
 end
 
-class Delayed::PerformableMethod
-  # serialize to YAML
-  def encode_with(coder)
-    coder.map = {
-      "object" => object,
-      "method_name" => method_name,
-      "args" => args
-    }
+module Delayed
+  class PerformableMethod
+    # serialize to YAML
+    def encode_with(coder)
+      coder.map = {
+        'object' => object,
+        'method_name' => method_name,
+        'args' => args
+      }
+    end
   end
 end
 
 module Psych
   module Visitors
-    class YAMLTree
-      def visit_Class(klass)
-        @emitter.scalar klass.name, nil, '!ruby/class', false, false, Nodes::Scalar::SINGLE_QUOTED
-      end
-    end
-
     class ToRuby
-      def visit_Psych_Nodes_Scalar(o)
-        @st[o.anchor] = o.value if o.anchor
-
-        if klass = Psych.load_tags[o.tag]
-          instance = klass.allocate
-
-          if instance.respond_to?(:init_with)
-            coder = Psych::Coder.new(o.tag)
-            coder.scalar = o.value
-            instance.init_with coder
-          end
-
-          return instance
-        end
-
-        return o.value if o.quoted
-        return @ss.tokenize(o.value) unless o.tag
-
-        case o.tag
-        when '!binary', 'tag:yaml.org,2002:binary'
-          o.value.unpack('m').first
-        when '!str', 'tag:yaml.org,2002:str'
-          o.value
-        when "!ruby/object:DateTime"
-          require 'date'
-          @ss.parse_time(o.value).to_datetime
-        when "!ruby/object:Complex"
-          Complex(o.value)
-        when "!ruby/object:Rational"
-          Rational(o.value)
-        when "!ruby/class", "!ruby/module"
-          resolve_class o.value
-        when "tag:yaml.org,2002:float", "!float"
-          Float(@ss.tokenize(o.value))
-        when "!ruby/regexp"
-          o.value =~ /^\/(.*)\/([mixn]*)$/
-          source  = $1
-          options = 0
-          lang    = nil
-          ($2 || '').split('').each do |option|
-            case option
-            when 'x' then options |= Regexp::EXTENDED
-            when 'i' then options |= Regexp::IGNORECASE
-            when 'm' then options |= Regexp::MULTILINE
-            when 'n' then options |= Regexp::NOENCODING
-            else lang = option
-            end
-          end
-          Regexp.new(*[source, options, lang].compact)
-        when "!ruby/range"
-          args = o.value.split(/([.]{2,3})/, 2).map { |s|
-            accept Nodes::Scalar.new(s)
-          }
-          args.push(args.delete_at(1) == '...')
-          Range.new(*args)
-        when /^!ruby\/sym(bol)?:?(.*)?$/
-          o.value.to_sym
-        else
-          @ss.tokenize o.value
-        end
-      end
-
-      def visit_Psych_Nodes_Mapping_with_class(object)
+      def visit_Psych_Nodes_Mapping_with_class(object) # rubocop:disable PerceivedComplexity, CyclomaticComplexity, MethodName
         return revive(Psych.load_tags[object.tag], object) if Psych.load_tags[object.tag]
 
         case object.tag
         when /^!ruby\/ActiveRecord:(.+)$/
-          klass = resolve_class($1)
+          klass = resolve_class(Regexp.last_match[1])
           payload = Hash[*object.children.map { |c| accept c }]
-          id = payload["attributes"][klass.primary_key]
+          id = payload['attributes'][klass.primary_key]
+          id = id.value if defined?(ActiveRecord::Attribute) && id.is_a?(ActiveRecord::Attribute)
           begin
             klass.unscoped.find(id)
-          rescue ActiveRecord::RecordNotFound
-            raise Delayed::DeserializationError
+          rescue ActiveRecord::RecordNotFound => error
+            raise Delayed::DeserializationError, "ActiveRecord::RecordNotFound, class: #{klass}, primary key: #{id} (#{error.message})"
           end
         when /^!ruby\/Mongoid:(.+)$/
-          klass = resolve_class($1)
+          klass = resolve_class(Regexp.last_match[1])
           payload = Hash[*object.children.map { |c| accept c }]
+          id = payload['attributes']['_id']
           begin
-            klass.find(payload["attributes"]["_id"])
-          rescue Mongoid::Errors::DocumentNotFound
-            raise Delayed::DeserializationError
+            klass.find(id)
+          rescue Mongoid::Errors::DocumentNotFound => error
+            raise Delayed::DeserializationError, "Mongoid::Errors::DocumentNotFound, class: #{klass}, primary key: #{id} (#{error.message})"
           end
         when /^!ruby\/DataMapper:(.+)$/
-          klass = resolve_class($1)
+          klass = resolve_class(Regexp.last_match[1])
           payload = Hash[*object.children.map { |c| accept c }]
           begin
-            primary_keys = klass.properties.select { |p| p.key? }
+            primary_keys = klass.properties.select(&:key?)
             key_names = primary_keys.map { |p| p.name.to_s }
-            klass.get!(*key_names.map { |k| payload["attributes"][k] })
-          rescue DataMapper::ObjectNotFoundError
-            raise Delayed::DeserializationError
+            klass.get!(*key_names.map { |k| payload['attributes'][k] })
+          rescue DataMapper::ObjectNotFoundError => error
+            raise Delayed::DeserializationError, "DataMapper::ObjectNotFoundError, class: #{klass} (#{error.message})"
           end
         else
           visit_Psych_Nodes_Mapping_without_class(object)