+++ spec/delayed/backend/test.rb
@@ -25,12 +25,11 @@ module Delayed
           self.attempts = 0
           self.priority = 0
           self.id = (self.class.id += 1)
-          hash.each{|k,v| send(:"#{k}=", v)}
+          hash.each { |k, v| send(:"#{k}=", v) }
         end
 
-        @jobs = []
         def self.all
-          @jobs
+          @jobs ||= []
         end
 
         def self.count
@@ -47,29 +46,33 @@ module Delayed
           end
         end
 
-        def self.create!(*args); create(*args); end
+        def self.create!(*args)
+          create(*args)
+        end
 
         def self.clear_locks!(worker_name)
-          all.select{|j| j.locked_by == worker_name}.each {|j| j.locked_by = nil; j.locked_at = nil}
+          all.select { |j| j.locked_by == worker_name }.each do |j|
+            j.locked_by = nil
+            j.locked_at = nil
+          end
         end
 
         # Find a few candidate jobs to run (in case some immediately get locked by others).
-        def self.find_available(worker_name, limit = 5, max_run_time = Worker.max_run_time)
+        def self.find_available(worker_name, limit = 5, max_run_time = Worker.max_run_time) # rubocop:disable CyclomaticComplexity, PerceivedComplexity
           jobs = all.select do |j|
             j.run_at <= db_time_now &&
-            (j.locked_at.nil? || j.locked_at < db_time_now - max_run_time || j.locked_by == worker_name) &&
-            !j.failed?
+              (j.locked_at.nil? || j.locked_at < db_time_now - max_run_time || j.locked_by == worker_name) &&
+              !j.failed?
           end
-
-          jobs = jobs.select{|j| Worker.queues.include?(j.queue)}   if Worker.queues.any?
-          jobs = jobs.select{|j| j.priority >= Worker.min_priority} if Worker.min_priority
-          jobs = jobs.select{|j| j.priority <= Worker.max_priority} if Worker.max_priority
-          jobs.sort_by{|j| [j.priority, j.run_at]}[0..limit-1]
+          jobs.select! { |j| j.priority <= Worker.max_priority } if Worker.max_priority
+          jobs.select! { |j| j.priority >= Worker.min_priority } if Worker.min_priority
+          jobs.select! { |j| Worker.queues.include?(j.queue) } if Worker.queues.any?
+          jobs.sort_by! { |j| [j.priority, j.run_at] }[0..limit - 1]
         end
 
         # Lock this job for this worker.
         # Returns true if we have the lock, false otherwise.
-        def lock_exclusively!(max_run_time, worker)
+        def lock_exclusively!(_max_run_time, worker)
           now = self.class.db_time_now
           if locked_by != worker
             # We don't own this job so we will update the locked_by name and the locked_at
@@ -77,7 +80,7 @@ module Delayed
             self.locked_by = worker
           end
 
-          return true
+          true
         end
 
         def self.db_time_now
@@ -85,7 +88,7 @@ module Delayed
         end
 
         def update_attributes(attrs = {})
-          attrs.each{|k,v| send(:"#{k}=", v)}
+          attrs.each { |k, v| send(:"#{k}=", v) }
           save
         end
 
@@ -100,7 +103,9 @@ module Delayed
           true
         end
 
-        def save!; save; end
+        def save!
+          save
+        end
 
         def reload
           reset