+++ spec/helper.rb
@@ -1,3 +1,15 @@
+require 'simplecov'
+require 'coveralls'
+
+SimpleCov.formatters = [SimpleCov::Formatter::HTMLFormatter, Coveralls::SimpleCov::Formatter]
+
+SimpleCov.start do
+  add_filter '/spec/'
+  # Each version of ruby and version of rails test different things
+  # This should probably just be removed.
+  minimum_coverage(85.0)
+end
+
 require 'logger'
 require 'rspec'
 
@@ -8,17 +20,14 @@ require 'active_record'
 require 'delayed_job'
 require 'delayed/backend/shared_spec'
 
-require 'simplecov'
-require 'coveralls'
-
-SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
-  SimpleCov::Formatter::HTMLFormatter,
-  Coveralls::SimpleCov::Formatter
-]
-SimpleCov.start
-
 Delayed::Worker.logger = Logger.new('/tmp/dj.log')
 ENV['RAILS_ENV'] = 'test'
+
+module Rails
+  def self.root
+    '.'
+  end
+end
 
 Delayed::Worker.backend = :test
 
@@ -27,7 +36,6 @@ ActiveSupport::Dependencies.autoload_paths << File.dirname(__FILE__)
 
 # Add this to simulate Railtie initializer being executed
 ActionMailer::Base.extend(Delayed::DelayMail)
-
 
 # Used to test interactions between DJ and an ORM
 ActiveRecord::Base.establish_connection :adapter => 'sqlite3', :database => ':memory:'
@@ -43,8 +51,13 @@ end
 
 class Story < ActiveRecord::Base
   self.primary_key = 'story_id'
-  def tell; text; end
-  def whatever(n, _); tell*n; end
+  def tell
+    text
+  end
+
+  def whatever(n, _)
+    tell * n
+  end
   default_scope { where(:scoped => true) }
 
   handle_asynchronously :whatever