+++ spec/performable_method_spec.rb
@@ -1,46 +1,44 @@
 require 'helper'
 
 describe Delayed::PerformableMethod do
-  describe "perform" do
+  describe 'perform' do
     before do
-      @method = Delayed::PerformableMethod.new("foo", :count, ['o'])
+      @method = Delayed::PerformableMethod.new('foo', :count, ['o'])
     end
 
-    context "with the persisted record cannot be found" do
+    context 'with the persisted record cannot be found' do
       before do
         @method.object = nil
       end
 
-      it "does nothing if object is nil" do
-        expect{@method.perform}.not_to raise_error
+      it 'does nothing if object is nil' do
+        expect { @method.perform }.not_to raise_error
       end
     end
 
-    it "calls the method on the object" do
+    it 'calls the method on the object' do
       expect(@method.object).to receive(:count).with('o')
       @method.perform
     end
   end
 
   it "raises a NoMethodError if target method doesn't exist" do
-    expect {
+    expect do
       Delayed::PerformableMethod.new(Object, :method_that_does_not_exist, [])
-    }.to raise_error(NoMethodError)
+    end.to raise_error(NoMethodError)
   end
 
-  it "does not raise NoMethodError if target method is private" do
+  it 'does not raise NoMethodError if target method is private' do
     clazz = Class.new do
       def private_method
       end
       private :private_method
     end
-    expect {
-      Delayed::PerformableMethod.new(clazz.new, :private_method, [])
-    }.not_to raise_error
+    expect { Delayed::PerformableMethod.new(clazz.new, :private_method, []) }.not_to raise_error
   end
 
-  describe "hooks" do
-    %w(before after success).each do |hook|
+  describe 'hooks' do
+    %w[before after success].each do |hook|
       it "delegates #{hook} hook to object" do
         story = Story.create
         job = story.delay.tell
@@ -50,7 +48,7 @@ describe Delayed::PerformableMethod do
       end
     end
 
-    %w(before after success).each do |hook|
+    %w[before after success].each do |hook|
       it "delegates #{hook} hook to object" do
         story = Story.create
         job = story.delay.tell
@@ -60,28 +58,28 @@ describe Delayed::PerformableMethod do
       end
     end
 
-    it "delegates enqueue hook to object" do
+    it 'delegates enqueue hook to object' do
       story = Story.create
       expect(story).to receive(:enqueue).with(an_instance_of(Delayed::Job))
       story.delay.tell
     end
 
-    it "delegates error hook to object" do
+    it 'delegates error hook to object' do
       story = Story.create
       expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
       expect(story).to receive(:tell).and_raise(RuntimeError)
       expect { story.delay.tell.invoke_job }.to raise_error
     end
 
-    it "delegates error hook to object when delay_jobs = false" do
+    it 'delegates error hook to object when delay_jobs = false' do
       story = Story.create
       expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
       expect(story).to receive(:tell).and_raise(RuntimeError)
       expect { story.delay.tell.invoke_job }.to raise_error
     end
 
-    it "delegates failure hook to object" do
-      method = Delayed::PerformableMethod.new("object", :size, [])
+    it 'delegates failure hook to object' do
+      method = Delayed::PerformableMethod.new('object', :size, [])
       expect(method.object).to receive(:failure)
       method.failure
     end
@@ -95,7 +93,7 @@ describe Delayed::PerformableMethod do
         Delayed::Worker.delay_jobs = true
       end
 
-      %w(before after success).each do |hook|
+      %w[before after success].each do |hook|
         it "delegates #{hook} hook to object" do
           story = Story.create
           expect(story).to receive(hook).with(an_instance_of(Delayed::Job))
@@ -103,7 +101,7 @@ describe Delayed::PerformableMethod do
         end
       end
 
-      %w(before after success).each do |hook|
+      %w[before after success].each do |hook|
         it "delegates #{hook} hook to object" do
           story = Story.create
           expect(story).to receive(hook).with(an_instance_of(Delayed::Job))
@@ -111,23 +109,23 @@ describe Delayed::PerformableMethod do
         end
       end
 
-      it "delegates error hook to object" do
+      it 'delegates error hook to object' do
         story = Story.create
         expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
         expect(story).to receive(:tell).and_raise(RuntimeError)
         expect { story.delay.tell }.to raise_error
       end
 
-      it "delegates error hook to object when delay_jobs = false" do
+      it 'delegates error hook to object when delay_jobs = false' do
         story = Story.create
         expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
         expect(story).to receive(:tell).and_raise(RuntimeError)
         expect { story.delay.tell }.to raise_error
       end
 
-      it "delegates failure hook to object when delay_jobs = false" do
+      it 'delegates failure hook to object when delay_jobs = false' do
         Delayed::Worker.delay_jobs = false
-        method = Delayed::PerformableMethod.new("object", :size, [])
+        method = Delayed::PerformableMethod.new('object', :size, [])
         expect(method.object).to receive(:failure)
         method.failure
       end